requirements/local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/local.txt requirements/base.in requirements/local.in
#
decorator==4.0.10         # via ipython, traitlets
django-codemirror2==0.2
django-extensions==1.7.4
django==1.10.3
factory-boy==2.7.0
fake-factory==0.7.2       # via factory-boy
ipython-genutils==0.1.0   # via traitlets
ipython==5.1.0
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.9     # via ipython
psycopg2==2.6.2
ptyprocess==0.5.1         # via pexpect
pygments==2.1.3           # via ipython
python-dateutil==2.6.0    # via fake-factory
pytz==2016.7
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-extensions, fake-factory, prompt-toolkit, python-dateutil, traitlets
tblib==1.3.0
traitlets==4.3.1          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
Werkzeug==0.11.11

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipython
